Add tests for DeepVariant scatter and merge
The test runs the test data with 2 scatters, and makes sure they are
present by checking the stdout from Cromwell. Also test to make sure
that the VCF output is still correct after scattering and merging.

For the test, we need the option to split the contigs (since the test
referene is so small), so a local modified version of
chunked-scatter.wdl that supports --split-contigs was added.