Make g.vcf output of the pipeline optional
This was done because generating g.vcf files using DeepVariant is very
slow (as in, it takes days when the variant calling takes hours).

When running the pipeline using GATK4 (the default), the g.vcf files are
always generated, but they will only be part of the pipeline output when
generateGVCF is set to true.